+++ src/Codeception/Lib/InnerBrowser.php
@@ -67,16 +67,6 @@ class InnerBrowser extends Module implements Web
 
     public function amOnPage($page)
     {
-        // use absolute url
-        if ((strpos($page, 'http://') === 0) or (strpos($page, 'https://') === 0)) {
-            $url = parse_url($page);
-            if (isset($url['host']) and isset($url['scheme'])) {
-                $host = $url['scheme'].'://'.$url['host'];
-                $this->_reconfigure(['url' => $host]);
-                $page = substr($page, strlen($host));
-                $this->debugSection('Host', $host);
-            }
-        }
         $this->crawler = $this->client->request('GET', $page);
         $this->forms = [];
         $this->debugResponse();