+++ src/Codeception/Module/PhpBrowser.php
@@ -2,6 +2,7 @@
 
 namespace Codeception\Module;
 
+use Codeception\Exception\TestRuntime;
 use Codeception\Lib\Connector\Guzzle;
 use Codeception\Lib\InnerBrowser;
 use Codeception\Lib\Interfaces\MultiSession;
@@ -103,7 +104,20 @@ class PhpBrowser extends InnerBrowser implements Remote, MultiSession
     public function amHttpAuthenticated($username, $password)
     {
         $this->client->setAuth($username, $password);
-    }    
+    }
+
+    public function amOnUrl($url)
+    {
+        $urlParts = parse_url($url);
+        if (!isset($urlParts['host']) or !isset($urlParts['scheme'])) {
+            throw new TestRuntime("Wrong URL passes, host and scheme not set");
+        }
+        $host = $urlParts['scheme'].'://'.$urlParts['host'];
+        $this->_reconfigure(['url' => $host]);
+        $page = substr($url, strlen($host));
+        $this->debugSection('Host', $host);
+        $this->amOnPage($page);
+    }
 
     public function amOnSubdomain($subdomain)
     {